+++ components/Footers/AuthFooter/AuthFooter.tsx
