+++ components/Footers/NotAuthFooter/NotAuthFooter.tsx
