+++ components/Forms/FormContainer/FormContainer.tsx
@@ -1,0 +1,6 @@
+import "./FormContainer.css";
+
+// adds default styles for form if we need to
+export function FormContainer({ children }: ChildrenProps) {
+  return <div className="form_container">{children}</div>;
+}
