+++ components/Loaders/Loader/Loader.tsx
@@ -1,9 +1,9 @@
 import "./Loader.css";
 
-export function Loader({ size }) {
+export function Loader({ size }: LoaderProps) {
   return (
     <div className="loader-overlay" style={{ fontSize: size }}>
       <span className="loader" />
     </div>
-  )
+  );
 }
